% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immigration.R
\name{immigration}
\alias{immigration}
\title{Immigration}
\usage{
immigration(hood, emigrants, weights)
}
\arguments{
\item{hood}{coordinates of the neighborhood of emigrating patch}

\item{emigrants}{the number of individuals to distribute among the neighborhood}

\item{weights}{the weights of neighborhood patches for probablistic drawing}
}
\value{
returns neighborhood coordinates and the number of times they were drawn
}
\description{
Distributes emigrants (via sampling with replacement) into neighborhood patches given their weights
}
