% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_sim.R
\name{stochastic_sim}
\alias{stochastic_sim}
\title{Stochastic metacommunity simulation}
\usage{
stochastic_sim(
  initial_df,
  aij,
  delta,
  r = 3,
  K = 100,
  timesteps = 5,
  disp_rate = 0.1,
  nh_size = 1,
  nh = "vonNeumann",
  dd_emi = FALSE,
  torus = TRUE,
  extinction_halt = TRUE,
  patch_extinction = 0
)
}
\arguments{
\item{initial_df}{dataframe specifying the initial conditions with ID, xy coordinates, speices denisities}

\item{aij}{interspecific interaction strength (aii assumed to be 1)}

\item{delta}{aggregation parameter where 0 is uniform weighting, >0 high densities are weighted more (<0 = less weight)}

\item{r}{growth rate of species (assumed equivalent for all species)}

\item{K}{carrying capacity of the environment (competition matrix is scaled according to K)}

\item{timesteps}{length of time to simulate dynamics}

\item{disp_rate}{dispersal rate which can range from 0 (no dispersal) to 1 (maximum dispersal)}

\item{nh_size}{Manhattan distance specifying size of neighborhood}

\item{nh}{Type of neighborhood for dispersal to occur. Can be either 'Von Neummann' or 'Moore'}

\item{dd_emi}{If TRUE, emigration is density-dependent and follows type-II functional response}

\item{torus}{If TRUE, the lattice is wrapped around to remove edge effects}

\item{extinction_halt}{If TRUE, simulation halts when one species goes extinct before timesteps is reached}

\item{patch_extinction}{probability specifying the rate in which a given patch becomes totally extinct (default is 0)}
}
\description{
Pieces functions together and runs a stochastic simulation of metacommunity dynamics
}
