% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_freq.R
\name{weighted_freq}
\alias{weighted_freq}
\title{Weighted frequencies}
\usage{
weighted_freq(N, delta)
}
\arguments{
\item{N}{densities of neighborhood cells}

\item{delta}{coefficient determining strength of frequency-dependence (higher freq patches receive more weights)}
}
\value{
returns the weights (sums up to 1) of each cell
}
\description{
Weights neighborhood frequencies according to an exponential function
}
